Refactor Graph to interface and Graph class to BaseGraph.
Making an interface a base for graph instead of a class made flow of abstraction way easier. Now GraphFactory and GraphGenerator can return <ChildClass extends Graph> instead of Graph, which enables returned instance to maintain the specialized context needed of ChildClass.

For instance if SpecializedGraph.class is needed for particular task:

SpecializedGraph g = GraphFactory.getInstance(SpecializedGraph.class, ...);

this statement will return an instance of a class with full functionalities, not a useless base class instance.